+++ CHANGES.md
@@ -16,6 +16,8 @@
 
 - Implement support for saving 3D viewers in session files. [#130]
 
+- Fix issue with _update_data on base VisPy viewer. [#106]
+
 - Optimize the layout of options for the layer style editors to save space. [#120]
 
 - Added the ability to save static images of the 3D viewers. [#125]
@@ -28,6 +30,8 @@
 
 - Fix bug that caused all layers in the 3D scatter viewer to disappear when
   one layer was removed.
+
+- Add selection toolbar and icons for 3D viewers. [#88] [#92] 
 
 0.2 (2015-03-11)
 ----------------